refactor: move axios to apis folder
