+++ src/app/pages/SignUp/helpers.ts
@@ -6,7 +6,7 @@ import { useMutation } from '@tanstack/react-query';
 import { zodResolver } from '@hookform/resolvers/zod';
 
 import { routes } from 'app/routes';
-import { axiosClient } from 'utils/axios';
+import { axiosClient } from 'apis/axios';
 import { queryKeys } from 'constants/queryKeys';
 
 import { messages } from './messages';
